tools/pixcel/pyproject.toml
[tool.poetry]
name = "pixcel"
version = "0.1.0"
description = ""
authors = ["nznznz42 <coolcarrot420@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.3.0"
openpyxl = "^3.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
